add cli

+++ Core/Inc/fs_logger.h
@@ -21,7 +21,7 @@ uint8_t FS_IsMounted(void);
 uint8_t FS_IsFormatted(void);
 
 // Event Logging
-void Log_AddEvent(uint16_t duration_ms, uint16_t rpm_val, uint8_t lane);
+void Log_AddEvent(uint32_t duration_ms, uint16_t rpm_val, uint8_t lane);
 void Log_FlushToCSV(void);
 void Log_StageEvents(void);
 void Log_ReadCSV(void);
